livekit-agent/requirements.txt
livekit-agents>=0.11.0
livekit-plugins-openai>=0.11.0
livekit-plugins-deepgram>=0.11.0
livekit-plugins-silero>=0.11.0
livekit-plugins-cartesia>=0.11.0
python-dotenv>=1.0.0
supabase>=2.0.0
